Move option_manager into unity_core to avoid layering violation
