+++ src/unity/lib/CMakeLists.txt
@@ -7,6 +7,8 @@ make_library(unity_core
     api/function_closure_info.cpp
     extensions/model_base.cpp
     extensions/ml_model.cpp
+    extensions/option_manager.cpp
+    extensions/option_info.cpp
     variant_converter.cpp
     variant.cpp
     unity_global.cpp
