+++ src/unity/lib/extensions/option_info.cpp
@@ -3,7 +3,8 @@
  * Use of this source code is governed by a BSD-3-clause license that can
  * be found in the LICENSE.txt file or at https://opensource.org/licenses/BSD-3-Clause
  */
-#include <unity/toolkits/options/option_info.hpp>
+#include <unity/lib/extensions/option_info.hpp>
+
 #include <boost/lexical_cast.hpp>
 
 namespace turi { namespace option_handling {
