+++ src/unity/lib/extensions/option_manager.cpp
@@ -3,7 +3,8 @@
  * Use of this source code is governed by a BSD-3-clause license that can
  * be found in the LICENSE.txt file or at https://opensource.org/licenses/BSD-3-Clause
  */
-#include <unity/toolkits/options/option_manager.hpp> 
+#include <unity/lib/extensions/option_manager.hpp>
+
 #include <boost/algorithm/string.hpp> 
 
 namespace turi {
