+++ src/unity/lib/extensions/option_manager.hpp
@@ -10,7 +10,7 @@
 #include <vector> 
 #include <map>
 
-#include <unity/toolkits/options/option_info.hpp> 
+#include <unity/lib/extensions/option_info.hpp>
 
 namespace turi {
