+++ src/unity/toolkits/recsys/models/popularity.cpp
@@ -3,9 +3,9 @@
  * Use of this source code is governed by a BSD-3-clause license that can
  * be found in the LICENSE.txt file or at https://opensource.org/licenses/BSD-3-Clause
  */
+#include <unity/lib/extensions/option_manager.hpp>
 #include <unity/lib/unity_sframe.hpp>
 #include <unity/toolkits/recsys/models/popularity.hpp>
-#include <unity/toolkits/options/option_manager.hpp> 
 #include <unity/toolkits/ml_data_2/metadata.hpp>
 #include <unity/toolkits/ml_data_2/ml_data.hpp>
 #include <unity/toolkits/ml_data_2/ml_data_iterators.hpp>
